refactor: extract select to separate files
