--- src/components/organisms/CompareModal/ResourceSetSelector.styled.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const ResourceSetSelectorDiv = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 9px 12px;
-  border-radius: 2;
-  background-color: #31393c;
-  margin-bottom: 16px;
-`;
