+++ src/components/organisms/CompareModal/ResourceSetSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './ResourceSetSelector';
